apply dialog in delete user, and loading datalist of contact

+++ diary-proyect/src/app/module/pages/contact/pages/list-contact/list-contact.component.ts
@@ -3,7 +3,8 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { ContactService } from 'src/app/module/services/contact.service';
 import { DialogConfirmationService } from '../../../../components/service/dialog-confirmation.service';
 import { DialogLoadingService } from 'src/app/module/components/service/dialog-loading.service';
-import { Title } from '@angular/platform-browser';
+import { DialogConfirmationComponent } from '../../../../components/dialog-confirmation/dialog-confirmation.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-list-contact',
@@ -27,8 +28,8 @@ export class ListContactComponent implements OnInit {
   constructor(
     private ContactService: ContactService,
     private snackbar: MatSnackBar,
-    private dialogConfirmation:DialogConfirmationService,
-    private dialogLoading:DialogLoadingService
+    private dialogLoading: DialogLoadingService,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -36,7 +37,7 @@ export class ListContactComponent implements OnInit {
   }
 
   getAllContacts(offset: number, searchTerm: string) {
-    this.dialogLoading.show('Cargando', 'Espere por favor...Cargando');
+    this.dialogLoading.show('Espere por favor, Cargando...', 'Consultando contactos');
 
     //calculate offset
     this.offset = (offset - 1) * this.limit;
@@ -51,56 +52,48 @@ export class ListContactComponent implements OnInit {
 
     this.ContactService.getAllContacts(data).subscribe({
       next: (response) => {
-        if(response.succeed){
-            
-      
+        if (response.succeed) {
+          console.log(response);
+          console.log((this.allContact = response.result));
+          console.log((this.allContact = this.allContact + this.countContact));
 
-      
+          this.allContact = response.result.count;
+          this.allContact = this.allContact + this.countContact;
+          localStorage.setItem('contact', this.allContact.toString());
 
+          console.log('offset', this.offset);
+          console.log('searchTerm', this.searchTerm);
+          console.log('limit', this.limit);
+          console.log('allContact', this.allContact);
+          console.log('currentPage', this.currentPage);
+          console.log('totalPages', this.totalPages);
 
-        console.log(response);
-        console.log((this.allContact = response.result));
-        console.log((this.allContact = this.allContact + this.countContact));
+          console.log((this.dataTable = response.result.count));
+          const data = response.result.list;
+          this.dataTable = data;
 
-        this.allContact = response.result.count;
-        this.allContact = this.allContact + this.countContact;
-        localStorage.setItem('contact', this.allContact.toString());
+          // console.log(data.length + this.offset, "data length");
+          // console.log(this.limit + this.limit - this.allContact);
+          // console.log((data.length + this.offset) / this.limit + this.limit - this.allContact);
 
-        console.log('offset', this.offset);
-        console.log('searchTerm', this.searchTerm);
-        console.log('limit', this.limit);
-        console.log('allContact', this.allContact);
-        console.log('currentPage', this.currentPage);
-        console.log('totalPages', this.totalPages);
-
-        console.log((this.dataTable = response.result.count));
-        const data = response.result.list;
-        this.dataTable = data;
-
-        // console.log(data.length + this.offset, "data length");
-        // console.log(this.limit + this.limit - this.allContact);
-        // console.log((data.length + this.offset) / this.limit + this.limit - this.allContact);
-
-        const end =
-          (data.length + this.offset) / this.limit +
-          this.limit -
-          this.allContact;
+          const end =
+            (data.length + this.offset) / this.limit +
+            this.limit -
+            this.allContact;
           this.dataTable = data.result.list.splice(0, this.limit - end);
 
-        this.totalPages = Math.ceil(this.allContact / this.limit);
-        this.currentPage = this.offset / this.limit + 1;
-        this.allContactChance.emit(this.allContact);
-
-      }
+          this.totalPages = Math.ceil(this.allContact / this.limit);
+          this.currentPage = this.offset / this.limit + 1;
+          this.allContactChance.emit(this.allContact);
+        }
       },
-      error : (err)=>{
+      error: (err) => {
         console.log(err);
-
       },
 
       complete: () => {
         this.dialogLoading.finish();
-      }
+      },
     });
   }
   pageChanged(event: any): void {
@@ -108,33 +101,54 @@ export class ListContactComponent implements OnInit {
     this.currentPage = event.pageIndex + 1;
     console.log(event.pageIndex + 1);
     this.getAllContacts(this.allContact, this.searchTerm);
-    
   }
-  deletContactId(id: number) {
-    this.ContactService.deleteContact(id).subscribe({
-      next: (response) => {
-        console.log(response);
-        this.countContact--;
-
-        this.snackbar.open(response.message, 'Aceptar', {
-          duration: 1000,
-        });
-
-        this.getAllContacts(this.offset, this.searchTerm);
-      },
-      error: (err) => {
-        this.snackbar.open(err.error.message, 'Aceptar', {
-          duration: 1000,
-        });
-        console.log(err);
-      },
-
-      //finish the next and error...
-      complete: () => {
-        this.getAllContacts(1, '');
-        this.allContactChance.emit(this.allContact);
+  deleteContactId(id: number) {
+    const dialog = this.dialog.open(DialogConfirmationComponent, {
+      width: '450px',
+      disableClose: true,
+      autoFocus: true,
+      data: {
+        title: 'Eliminar contacto',
+        message: '¿Estas seguro de eliminar este contacto?',
       },
     });
-  }
 
+
+    if (dialog) {
+
+      dialog.afterClosed().subscribe((confirmation) => {
+        if (confirmation) {
+          this.ContactService.deleteContact(id).subscribe({
+            next: (response) => {
+              if (response.succeed) {
+                console.log(response);
+                this.countContact--;
+  
+                this.dialogLoading.show( 'Cargando', '  Espere por favor...');
+                this.snackbar.open(response.message, 'Aceptar', {
+                  duration: 1000,
+                });
+  
+                this.getAllContacts(this.offset, this.searchTerm);
+              }
+            },
+            error: (err) => {
+              this.snackbar.open(err.error.message, 'Aceptar', {
+                duration: 1000,
+              });
+              console.log(err);
+            },
+  
+            //finish the next and error...
+            complete: () => {
+              this.getAllContacts(1, '');
+              this.allContactChance.emit(this.allContact);
+              this.dialogLoading.finish();
+            },
+          });
+        }
+      });
+    }
+
+  }
 }
